Functions Lab
created a function that returns the distance from Scubers HQ in blocks from the customers pick up location. Then created a function that used the first function as a call back to find the distance that would be travelled in feet. The third function finds the the distance in feet from to random points in a city. the fourth function calculates the fare price. ever foot travelled is 2 cents. However, in the fourth function the first 400 feet are free.